zistino_backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-filter==23.5
django-extensions==3.2.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
openpyxl==3.1.2
